pyproject.toml
[project]
name            = "ml-dev-kit"
version         = "0.0.1"
requires-python = ">=3.11"

authors = [{ name = "tonnetonne814", email = "theobroma814@gmail.com" }]

dependencies = [
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.1.post0",
    "optuna>=4.3.0",
    "pytest>=8.3.5",
    "rich>=14.0.0",
    "rootutils>=1.0.7",
    "sh>=2.2.2",
    "tensorboard>=2.19.0",
    "wandb>=0.19.10",
]

# specific dependencies
[project.optional-dependencies]
cpu = [
  "torch==2.7.0",
  "torchvision==0.22.0",
  "torchaudio==2.7.0",
]
cu128 = [
  "torch==2.7.0",
  "torchvision==0.22.0",
  "torchaudio==2.7.0",
]
cu126 = [
  "torch==2.6.0",
  "torchvision==0.21.0",
  "torchaudio==2.6.0",
]
cu118 = [
  "torch==2.5.1",
  "torchvision==0.20.1",
  "torchaudio==2.5.1",
]

# describe conflicts
[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cu128" },
    { extra = "cu126" },
    { extra = "cu118" },
  ],
]

# index mapping
[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu118", extra = "cu118" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu128", extra = "cu128" },
  { index = "pytorch-cu126", extra = "cu126" },
  { index = "pytorch-cu118", extra = "cu118" },
]

# index info
[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
